Move Mso code into a separate file
In prep for refactoring it to a flag-backed option.
